requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.1
django-environ>=0.11
requests>=2.32
cryptography>=42.0
gunicorn>=22.0
whitenoise>=6.6
django-cors-headers>=4.3
pytest>=8.0
pytest-django>=4.8
responses>=0.25
factory-boy>=3.3
sentry-sdk>=2.0
